@@ -1,10 +1,10 @@
 <script lang="ts">
-  import type { IHostel } from '$types/Amenity.types';
   import MapsSelector from '$utils/MapsSelector';
-  import copyToClipboard from '$utils/CopyToClipboard';
   import WHATSAPP from '$images/logo-whatsapp.png';
+  import type { IHostel } from '$types/Amenity.types';
+  import copyToClipboard from '$utils/CopyToClipboard';
 
-  export const data = $$props as IHostel;
+  export const data: IHostel = $props();
 </script>
 
 <div class="AmenityCard">
@@ -22,10 +22,10 @@
           href={'tel:' + data.phone}
           title="Copy phone number"
           class="CopyButton AmenityCard__separator--content"
-          on:keydown={async () => await copyToClipboard(data.phone)}
-          on:keyup={async () => await copyToClipboard(data.phone)}
-          on:keypress={async () => await copyToClipboard(data.phone)}
-          on:click={async () => await copyToClipboard(data.phone)}
+          onkeydown={async () => await copyToClipboard(data.phone)}
+          onkeyup={async () => await copyToClipboard(data.phone)}
+          onkeypress={async () => await copyToClipboard(data.phone)}
+          onclick={async () => await copyToClipboard(data.phone)}
         >
           {data.phone}
         </a>
@@ -42,10 +42,10 @@
           href={'mail:' + data.mail}
           title="Copy phone number"
           class="CopyButton AmenityCard__separator--content"
-          on:keydown={async () => await copyToClipboard(data.mail)}
-          on:keyup={async () => await copyToClipboard(data.mail)}
-          on:keypress={async () => await copyToClipboard(data.mail)}
-          on:click={async () => await copyToClipboard(data.mail)}
+          onkeydown={async () => await copyToClipboard(data.mail)}
+          onkeyup={async () => await copyToClipboard(data.mail)}
+          onkeypress={async () => await copyToClipboard(data.mail)}
+          onclick={async () => await copyToClipboard(data.mail)}
         >
           {data.mail}
         </a>
