@@ -1,7 +1,12 @@
 <script lang="ts">
   import type { PageData } from './$types';
   import copyToClipboard from '$utils/CopyToClipboard';
-  export let data: PageData;
+
+  let {
+    data
+  }: {
+    data: PageData;
+  } = $props();
 </script>
 
 <svelte:head>
@@ -20,10 +25,10 @@
         <button
           class="CopyButton"
           data-icon={String.fromCharCode(57688)}
-          on:keydown={async () => await copyToClipboard(club.email)}
-          on:keyup={async () => await copyToClipboard(club.email)}
-          on:keypress={async () => await copyToClipboard(club.email)}
-          on:click={async () => await copyToClipboard(club.email)}
+          onkeydown={async () => await copyToClipboard(club.email)}
+          onkeyup={async () => await copyToClipboard(club.email)}
+          onkeypress={async () => await copyToClipboard(club.email)}
+          onclick={async () => await copyToClipboard(club.email)}
           title="Click to copy email"
         >
           {club.email}
@@ -60,5 +65,5 @@
 </ul>
 
 <style lang="scss">
-  @import '../ClubStyle.scss';
+  @forward '../ClubStyle.scss';
 </style>
