@@ -1,32 +1,36 @@
-import adapter from '@sveltejs/adapter-auto';
-import preprocess from 'svelte-preprocess';
 import autoprefixer from 'autoprefixer';
+import adapter from '@sveltejs/adapter-auto';
+import { sveltePreprocess } from 'svelte-preprocess';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-  preprocess: preprocess({
+  preprocess: sveltePreprocess({
     scss: {
       prependData: `
-				@use "src/styles/root/_variables" as *;
-				@import "src/styles/root/_mixins.scss";
+        // This is only for <style> tags in Svelte files
+        // So we need to import the variables and mixins in every other scss file
+        @use "src/styles/root/_variables" as *;
+        @use "src/styles/root/_mixins.scss" as *;
 			`
     },
     postcss: {
       plugins: [autoprefixer()]
     }
   }),
-
   kit: {
     adapter: adapter(),
     alias: {
-      $components: 'src/lib/components',
-      $styles: 'src/styles/routes',
-      $data: 'src/lib/data',
-      $utils: 'src/lib/utils',
-      $types: 'src/lib/types',
+      $data: 'src/lib/data/*',
+      $utils: 'src/lib/utils/*',
+      $types: 'src/lib/types/*',
       $stores: 'src/lib/stores/*',
-      $images: 'src/lib/images/*'
+      $images: 'src/lib/images/*',
+      $styles: 'src/styles/routes/*',
+      $components: 'src/lib/components/*',
     }
+  },
+  compilerOptions: {
+    runes: true
   }
 };
 
